refactor: leave the interfaces in their dedicated folder.
This allow easy automatic import of the interfaces.
